[UPDATE]Successful test file separation.
