+++ CuRendCore/CuRendCore_Test/entry.cpp
@@ -1,0 +1,7 @@
+#include "pch.h"
+
+int main(int argc, char **argv)
+{
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
